Add simple string interpolater
Although there are quite a few string interpolation packages on Hackage,
they all come with a hefty set of dependencies (which can matter quite a
lot when executing a script with a Nix shebang when dependencies are
fetched there and then). This has no dependencies (besides `base`).

I've found it can really help tidy up `shh` scripts, it's also generic
over any type with an `IsString` instance (of course the usual nonsense
regarding the `ByteString` instance applies).

Comes with raw and unindenting versions.